Change condition from tripLocation to imageid
This condition should have been for checking if the imageid is populated on an update, tripLocation should not have been used for this condition, has been corrected. This check should always pass as imageid should not ever be undefined but it is still good to have for clarity and in case something goes wrong.